Abstract some parts of ControlPlane
This will simplify other implementations.
